+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     - `math::FaceMap` now implements `Exhaust`.
     - `math::GridAab::intersection_box()` is a new definition of box intersection which returns a bigger box in some zero-volume/flat cases.
     - `math::GridAab::union_cubes()` is a new definition of union which returns a smaller box in zero-volume/flat cases.
+    - `math::GridAab::to_free()` is another name for `Aab::from()`, useful for method chaining and not importing `Aab`.
     - `math::GridSize` type alias for `euclid::Size3D<GridCoordinate, Cube>`.
     - `math::Vol::subdivide()` splits a `Vol`'s bounds and data into parts that can be processed in parallel.